VFB-419: moved common DeliveryInstructionsCard to /common/formSections
